Worked on the logical architecture of the application
I did the logical architecture of the application. I made the function that sends the project id as a url parameter. After that I get it using "usePrams()" on home page and then I pass the obtained url to the navigation bar. By passing the project id in the path link to each page, I make sure I will be able to access it everywhere