One of the last commits for this project.
I fixed sine small errors and bugs, as well as some styling on the page. I fixed the project component, and made it to pass the project id to the other pages with "navigate". I made the two "create-new" pages to send the proper information to the data base. I made the function, that on click on applicant's profile item or card, transfers you to the applicant in profile page, as well as it sends the data to the page dynamically