Small changes
I did a lot of small changes on the design. I fixed a solid amount of small design and layout bugs, that were to be fixed. I assigned a "ProjectId" to the reminders as well. Now the reminders are also being filtered accordingly the project you are currently in. I made the document fetching from the firebase with onSnapshot instead of getDoc, because it is faster and as well watches for changes in the documents, when there is a change, it updates the doc on the client side.